(refs #537) Rename bless to fieldify
