+++ src/web/forms/fields/index.js
@@ -34,25 +34,25 @@ import { Diagnoses } from './Diagnoses';
 import { MultiInput } from './MultiInput';
 import { SubformList, SubformListComponent, ReadonlySubformList } from './SubformList';
 import { Block } from './Block';
-import bless from './common/bless';
+import fieldify from './common/fieldify';
 
 const fieldComponentList = {
-  textinput: bless(TextInput),
-  textunitinput: bless(TextUnitInput),
-  textarea: bless(TextArea),
-  radio: bless(RadioGroup),
-  select: bless(Select),
-  attachment: bless(Attachment),
-  attachmentinput: bless(AttachmentInput),
-  attachmentviewer: bless(AttachmentViewer),
+  textinput: fieldify(TextInput),
+  textunitinput: fieldify(TextUnitInput),
+  textarea: fieldify(TextArea),
+  radio: fieldify(RadioGroup),
+  select: fieldify(Select),
+  attachment: fieldify(Attachment),
+  attachmentinput: fieldify(AttachmentInput),
+  attachmentviewer: fieldify(AttachmentViewer),
   accordion: Accordion,
-  check: bless(Checkbox),
-  checkgroup: bless(CheckGroup),
-  autocalc: bless(AutoCalc),
+  check: fieldify(Checkbox),
+  checkgroup: fieldify(CheckGroup),
+  autocalc: fieldify(AutoCalc),
   guide: GuideTools,
   diagnoses: Diagnoses,
-  multiinput: bless(MultiInput),
-  subformlist: bless(SubformList),
+  multiinput: fieldify(MultiInput),
+  subformlist: fieldify(SubformList),
   block: Block,
 };
